Refactor crate: Merge `mod fd` into `mod stdio`
`mod fd` only contains one single function `dup` and is only used in
`mod stdio`.
